调整了osnr返回值，现在会给出建议了

+++ src/main/java/com/otn/controller/LinkTypeController.java
@@ -88,7 +88,8 @@ public class LinkTypeController {
     @RequestMapping(value = "/{versionId}/{linkType}/{length}", method = RequestMethod.GET)
     @ResponseStatus(HttpStatus.OK)
     @VersionCheckException
-    public Loss calculateLoss(@PathVariable Long versionId, @PathVariable("linkType") String linkType, @PathVariable Long length) {
+    public Loss calculateLoss(@PathVariable Long versionId, @PathVariable("linkType") String linkType, @PathVariable
+            Double length) {
         return new Loss(linkTypeService.calculateLoss(versionId, linkType,length));
     }
     private class Loss {
